requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
requests==2.33.1
urllib3==2.5.0
beautifulsoup4==4.14.3
pydantic==2.13.3
numpy==2.4.4
pillow==12.2.0
